Fix unresponsiveness of Django server when the docker-make-chris_dev.sh script is run with -i flag.
Fix unresponsiveness of Django server whenthe docker-make-chris_dev.sh script is run with -i flag

Fix bug in which the raw value for a plugin parameter from a request was being passed to the plugin instance manager. Instead the serializer validated value is now passed.

Remove swift container creation code from tests and add it to the Django settings