Fixed CUBE minikube backend deployment
The issue raised which showed 'connection refused' was because the
chris_dev remote container and pfcon kubernetes pod were in different
networks. I solved this issue by adding minikube as an external network
in the docker-compose file, and the only way communication could be
ensured was using minikube's external IP as the HOSTIP. Note:This was
only tested with minikube having a docker driver, with other drivers
such as hyperkit, the outcome is not ensured.

A second issue after going past this was the error when the job(zip
file) was supposed to be submitted from pfcon to pman. These caused
tests in the ./make.sh bash script to fail.This issue was fixed by attaching a
role and binding the role to the default ServiceAccount in the default
Namespace. This issue doesn't exist in the prod setup because of the
availabily of these configurations. But it is helpful for those running
the development setup.